split existing tests into tests directory
